+++ test/directions_test.dart
@@ -1,16 +1,12 @@
-library google_maps_webservice.directions.test;
-
 import 'dart:async';
 import 'dart:convert';
 import 'package:google_maps_webservice/src/core.dart';
 import 'package:google_maps_webservice/src/directions.dart';
-import 'package:http/http.dart';
 import 'package:test/test.dart';
 
-Future<void> launch([Client client]) async {
+Future<void> main() async {
   final apiKey = 'MY_API_KEY';
-  GoogleMapsDirections directions =
-      GoogleMapsDirections(apiKey: apiKey, httpClient: client);
+  var directions = GoogleMapsDirections(apiKey: apiKey);
 
   tearDownAll(() {
     directions.dispose();
@@ -65,30 +61,37 @@ Future<void> launch([Client client]) async {
             directions.buildUrl(
                 origin: 'Toronto',
                 destination: 'Montreal',
-                avoid: RouteType.tolls),
+                avoids: [RouteType.tolls]),
             equals(
                 'https://maps.googleapis.com/maps/api/directions/json?origin=Toronto&destination=Montreal&avoid=tolls&key=$apiKey'));
         expect(
             directions.buildUrl(
                 origin: 'Toronto',
                 destination: 'Montreal',
-                avoid: RouteType.highways),
+                avoids: [RouteType.highways]),
             equals(
                 'https://maps.googleapis.com/maps/api/directions/json?origin=Toronto&destination=Montreal&avoid=highways&key=$apiKey'));
         expect(
             directions.buildUrl(
                 origin: 'Toronto',
                 destination: 'Montreal',
-                avoid: RouteType.indoor),
+                avoids: [RouteType.indoor]),
             equals(
                 'https://maps.googleapis.com/maps/api/directions/json?origin=Toronto&destination=Montreal&avoid=indoor&key=$apiKey'));
         expect(
             directions.buildUrl(
                 origin: 'Toronto',
                 destination: 'Montreal',
-                avoid: RouteType.ferries),
+                avoids: [RouteType.ferries]),
             equals(
                 'https://maps.googleapis.com/maps/api/directions/json?origin=Toronto&destination=Montreal&avoid=ferries&key=$apiKey'));
+        expect(
+            directions.buildUrl(
+                origin: 'Toronto',
+                destination: 'Montreal',
+                avoids: [RouteType.indoor, RouteType.ferries]),
+            equals(
+                'https://maps.googleapis.com/maps/api/directions/json?origin=Toronto&destination=Montreal&avoid=indoor|ferries&key=$apiKey'));
       });
 
       test('travel_mode', () {
@@ -123,7 +126,7 @@ Future<void> launch([Client client]) async {
       });
 
       test('departure_time', () {
-        int d = 1343641500;
+        var d = 1343641500;
         expect(
             directions.buildUrl(
                 origin: 'Toronto', destination: 'Montreal', departureTime: d),
@@ -138,8 +141,19 @@ Future<void> launch([Client client]) async {
                 'https://maps.googleapis.com/maps/api/directions/json?origin=Toronto&destination=Montreal&departure_time=$d&key=$apiKey'));
       });
 
+      test('departure_time with now', () {
+        expect(
+            directions.buildUrl(
+                origin: 'Toronto',
+                destination: 'Montreal',
+                departureTime: 'now'),
+            equals(
+              'https://maps.googleapis.com/maps/api/directions/json?origin=Toronto&destination=Montreal&departure_time=now&key=$apiKey',
+            ));
+      });
+
       test('arrival_time', () {
-        int d = 1343641500;
+        var d = 1343641500;
         expect(
             directions.buildUrl(
                 origin: 'Toronto', destination: 'Montreal', arrivalTime: d),
@@ -255,11 +269,25 @@ Future<void> launch([Client client]) async {
             equals(
                 'https://maps.googleapis.com/maps/api/directions/json?origin=Toronto&destination=Montreal&waypoints=optimize:true|Paris|42.2,21.3|place_id:ChIJ3S-JXmauEmsRUcIaWtf4MzE|enc:gfo}EtohhU:&key=$apiKey'));
       });
+
+      test('alternatives', () {
+        expect(
+            directions.buildUrl(
+                origin: 'Toronto', destination: 'Montreal', alternatives: true),
+            equals(
+                'https://maps.googleapis.com/maps/api/directions/json?origin=Toronto&destination=Montreal&alternatives=true&key=$apiKey'));
+        expect(
+            directions.buildUrl(
+                origin: 'Toronto',
+                destination: 'Montreal',
+                alternatives: false),
+            equals(
+                'https://maps.googleapis.com/maps/api/directions/json?origin=Toronto&destination=Montreal&alternatives=false&key=$apiKey'));
+      });
     });
 
     test('decode response', () {
-      DirectionsResponse response =
-          DirectionsResponse.fromJson(json.decode(_responseExample));
+      var response = DirectionsResponse.fromJson(json.decode(_responseExample));
 
       expect(response.isOkay, isTrue);
       expect(response.routes, hasLength(equals(1)));
@@ -437,5 +465,3 @@ final _responseExample = '''
   } ]
 }
 ''';
-
-Future<void> main() => launch();